+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Future]
 
 ### Added
 ### Changed
 ### Removed
 ### Fixed
+
+
+## [Unreleased]
+
+### Added
+* The client now saves the player/remote setting in the browser.
+* The noise toggle button now shows an animation when the noise is running.
 
 
 ## [0.1.0] - 2019-11-02